- smaller size for some icons (thanks to Maverick) - now rate_limiter_delete.php and site_bans_delete.php remove their lang keys and strings - the trigger for photos updates last_modified in user's blog posts and comments if the main photo changed so the cache for these can be regenerated with the new photo - phpmailer default encoding is now utf-8. Must check if this is ok. - added a cron to remove orphaned lk ids. - some changes in user_cache.class.php to accept non-cached content (for the last_online for now) Fixed issues: 2,12,6

+++ trunk/admin/processors/rate_limiter_delete.php
@@ -21,11 +21,20 @@ $qs_sep='';
 $topass=array();
 $rate_id=isset($_GET['rate_id']) ? (int)$_GET['rate_id'] : 0;
 
-$query="DELETE FROM `{$dbtable_prefix}rate_limiter` WHERE `rate_id`=$rate_id";
+$query="SELECT `fk_lk_id_error_message` FROM `{$dbtable_prefix}rate_limiter` WHERE `rate_id`=$rate_id";
 if (!($res=@mysql_query($query))) {trigger_error(mysql_error(),E_USER_ERROR);}
-regenerate_langstrings_array();
+if (mysql_num_rows($res)) {
+	$lk_id=mysql_result($res,0,0);
+	$query="DELETE FROM `{$dbtable_prefix}lang_strings` WHERE `fk_lk_id`=$lk_id";
+	if (!($res=@mysql_query($query))) {trigger_error(mysql_error(),E_USER_ERROR);}
+	$query="DELETE FROM `{$dbtable_prefix}lang_keys` WHERE `lk_id`=$lk_id";
+	if (!($res=@mysql_query($query))) {trigger_error(mysql_error(),E_USER_ERROR);}
+	$query="DELETE FROM `{$dbtable_prefix}rate_limiter` WHERE `rate_id`=$rate_id";
+	if (!($res=@mysql_query($query))) {trigger_error(mysql_error(),E_USER_ERROR);}
+	regenerate_langstrings_array();
 
-$topass['message']['type']=MESSAGE_INFO;
-$topass['message']['text']='Limit removed successfully.';
+	$topass['message']['type']=MESSAGE_INFO;
+	$topass['message']['text']='Limit removed successfully.';
+}
 
 redirect2page('admin/rate_limiter.php',$topass,$qs);
